Refactoring started, just merging master rn
