+++ frontend/src/Routing.js
@@ -5,7 +5,6 @@ import CreateEmployee from './employees/CreateEmployee'
 import UpdateEmployee from './employees/UpdateEmployee'
 import TableListOfWorkSheets from './worksheets/TableListOfWorkSheets'
 import TableListofPartners from './partners/TableListofPartners'
-import CreateWorkSheet from './worksheets/CreateWorkSheet'
 import UpdateWorkSheet from './worksheets/UpdateWorkSheet'
 import useCurrentUser from './hooks/useCurrentUser'
 
@@ -31,7 +30,7 @@ function Routing() {
           <TableListofPartners />
         </Route>
         <Route path="/worksheets/new">
-          <CreateWorkSheet />
+          <UpdateWorkSheet />
         </Route>
         <Route path="/worksheets/update/:id">
           <UpdateWorkSheet />
